Bug 668619 - slapd stops responding
https://bugzilla.redhat.com/show_bug.cgi?id=668619

Description: This patch contains 2 fixes:
1) I am moving the disconnect_server_nomutex into the section
where it adds the fd to the poll list - if we are going to
disconnect the connection due to timelimit, we should not add
the fd to the poll list - we should also remove the connection
from the active list. (by rmeggins@redhat.com)

2) If simple paged search result request hits timelimit,
the request is abandoned and the search result set allocated in
the backend module is released.  The release is executed by the
main thread, which is protected by the c_mutex (in Connection).
The search result is sent back to the client by a worker thread.
The worker thread had accessed the search result set without any
checking.  This patch adds the protection using c_mutex if the
search is simple paged.