Bug 710377 - Import with chain-on-update crashes ns-slapd
If you perform an import using ldif2db for a suffix that is configured
to chain-on-update, ns-slapd will crash due to a NULL pointer dereference.
The problem is that the chaining backends are not loaded when ns-slapd is
run in ldif2db mode.  This results in the chaining backend being NULL.

To fix this issue, we should return the local backend in the chain-on-update
distribution function when we are unable to find a backend by name.  We can
assume that this only occurs when an import is being done.  This should
allow one to use ldif2db to do an offline replica initialization when
chain-on-update is configured.