add support for ldif files with changetype: add
IPA config files use changetype: add to add new files.  This was not
supported by the DS ConfigFile directive or other LDIF processing.
Now, if we see that the entry has changetype: add, we process it as if
it had no changetype at all, which is to add the entry.  Remove
the changetype: from the entry before adding.
Reviewed by: nkinder (Thanks!)