Bug 709468 - RSA Authentication Server timeouts when using simple paged results on RHDS 8.2.
https://bugzilla.redhat.com/show_bug.cgi?id=709468
Resolves: bug 709468
Bug Description: RSA Authentication Server timeouts when using simple paged results on RHDS 8.2.
Reviewed by: nkinder, nhosoi (Thanks!)
Branch: master
Fix Description: When returning the last page of the last result set, in the
searchResultDone message, the paged results cookie is set to NULL.  This
means the client cannot issue another search using this result set/cookie,
which means the server can free/reset the simple paged resources
associated with this result set.  Use pagedresults_cleanup() to do this.
Since before it was assumed pagedresults_cleanup was called with the
connection lock, a new parameter tells it if it needs to lock or not.
Platforms tested: RHEL6 x86_64
Flag Day: no
Doc impact: no