Bug 725743 - Make memberOf use PRMonitor for it's operation lock
We currently use a mutex for the operation lock in the memberOf
plug-in.  This can cause a deadlock issue since the locks are not
re-entrant.  When memberOf performs an internal operation that
triggers another plug-in to perform an internal operation, a
deadlock will occur if this second internal operation triggers
memberOf again.

This patch switches memberOf to use PRMonitor instead, which is
re-entrant.
(cherry picked from commit 089971f454d768c0cff4af494db16bffa5260c6c)