Bug 728510 - Run dirsync after sending updates to AD
There is currently a window of time between sending changes to AD
and running a dirsync to get changes from AD that can cause problems.
We rely on having the GUID from AD for certain operations on entries,
such as deletes.  When a new entry is synced to AD, we don't get the
GUID back from AD until a dirsync is run, which is at a 5 minute
interval by default.  When changes that rely on DS having the GUID
are performed during this window, things can get out of sync.

This patch makes us run a dirsync immediately after sending updates
to AD for both the total and incremental update sync protocols.  This
minimizes the window where we do not have the GUID for an entry, thus
minimizing the chance of an update occuring in that window.