Bug 729717 - Fatal error messages when syncing deletes from AD
When we sync a delete from AD, we receive some fatal error messages
about failed searches for the deleted entry against AD.  This exposed
some operations that we shouldn't be doing, as well as a message being
logged at an incorrect log level.

We were attempting to enable a user entry in AD for a delete operation.
We should only try to enable a user entry that we just created in AD
(or recently added a password to).  We should not try to enable a user
for delete operations.

We also attempt to delete an entry in AD when processing a local delete,
even is the entry was not found in AD.  If the entry is not found in AD,
we should just skip sending the delete operation to AD.