Bug 611438 - Add Account Usability Control support
This adds support for the account usability request and response
controls that are used by Solaris clients.  This control allows
one to check if an account is usable without actually performing
a bind attempt as that entry.  The OpenDS documentation describes
the control in greater detail.

Various information about the password policy is returned in the
response control.  To allow the new plug-in to get the password
policy info (and to make password policy checks), I added some
new SLAPI functions around password policies.  I did this in such
a way that we can add support for password policy plug-ins in the
future if we expand upon this new API.  The password policy is
represented by an opaque Slapi_PWPolicy struct, which is currently
just our internal pwpolicy struct.  This will allow password policy
plug-ins to have their own struct that meets their needs if we add
support for constructor and destructor callbacks.

By default, we only allow the root DN to use this control, though
one can set an aci in the control entry under cn=features to allow
others to use the control.