Bug 732541 - Ignore error 32 when adding automember config
When you add a new automember config entry, we validate it at the
pre-op stage.  Part of this validation includes searching for child
regex rule entries.  We use the config entry DN as the search base,
but this fails when a new config entry is being added since it the
entry doesn't yet exist at the pre-op phase.  This results in the
new entry being considered invalid, which causes us to return an
unwilling to perform error that blocks the config entry from
being added.

We can safely ignore a LDAP_NO_SUCH_OBJECT error when attempting
to search for child regex rules.