Bug 733103 - large targetattr list with syntax errors cause server to crash or hang
https://bugzilla.redhat.com/show_bug.cgi?id=733103
Resolves: bug 733103
Bug Description: large targetattr list with syntax errors cause server to crash or hang
Reviewed by: nhosoi (Thanks!)
Branch: 389-ds-base-1.2.9
Fix Description: If we need to realloc the list to hold more target attributes,
the list is not NULL terminated.  If there are syntax errors in the aci, the
function returns with an unterminated list, and this will either cause the
loop in the free code to go off the end of the list or crash with a segfault.
The fix is to make sure the list is always NULL terminated.
Platforms tested: RHEL6 x86_64
Flag Day: no
Doc impact: no
(cherry picked from commit 91c6799b07ca48bce85482136ceb244b484fc91b)