Bug 732153 - subtree and user account lockout policies implemented?
https://bugzilla.redhat.com/show_bug.cgi?id=732153

Description: Upgradedbformat utility converts old dn format to
new one.  It's called at the in-place upgrade time.  The utility
failed to check if the normalized RDN value is in the entry's
attribute list.  This patch properly checks it.  If the value
does not exist, it adds the normalized value to the entry.