Bug 722292 - Entries in DS are not updated properly when using WinSync API
When an entry exists in a container entry (such as an OU) in AD, changes
will not be synced to DS when using a winsync API plug-in that flattens
the DN.

The problem is that the code that detects if a rename or move of an
entry in AD has occurred does not call the winsync API plug-in to get
the local DN.  This would cause the local entry to not be found, which
makes the modify fail.

This patch makes sure the winsync API plug-in is called when it tries
to detect if a rename was performed.  In addition, I found that we
need to re-fetch the newly renamed local entry before attempting to
apply any modifications to the local entry that we receive from AD.
Without doing this, the modify was failing since the old DN was being
used, resulting in an error 32.
(cherry picked from commit 402cd420ebb05aba2ebf80b61721578c2b06d9be)