Bug 735114 - renaming a managed entry does not update mepmanagedby
When a number of different plug-ins are being used that perform
internal operations to update entries, it is possible for the
managed entry plug-in to fail to update the managed entry.  In
particular, it has been found that renaming an origin entry can
trigger a plug-in other than the managed entry plug-in to rename
the managed entry.  This causes the managed entry plug-in to fail
when it attempts to update the managed entry.

This patch makes the managed entry plug-in check if the managed
entry has already been renamed when processing a MODRDN operation.
If it detects that the managed entry has already been renamed, it
will use the new DN to perform the managed entry updates.