Bug 590826 - Reloading database from ldif causes changelog to emit "data no longer matches" errors
https://bugzilla.redhat.com/show_bug.cgi?id=590826
Resolves: bug 590826
Bug Description: Reloading database from ldif causes changelog to emit "data no longer matches" errors
Reviewed by: nkinder, nhosoi (Thanks!)
Branch: master
Fix Description: When there are obsolete or decommissioned masters in the
RUV, this should not invalidate the changelog.  Instead, warn the user that
there are replicas in the database RUV that are not in the changelog
max RUV.  In this case, the CLEANRUV task can be used to remove the
obsolete masters from the database RUV.  I had to add a function to
generate a string representation of the replica RUVElement* for logging
purposes, and used that function elsewhere.  The new function
ruv_compare_ruv should be used instead of ruv_contains_ruv since it gives
more flexibility about logging and handling different cases.
Platforms tested: RHEL6 x86_64
Flag Day: no
Doc impact: no