Bug 739172 - Allow separate fractional attrs for incremental and total protocols
This patch adds the ability to have a separate fractional attribute
list for the incremental and total update replication protocols.

A new nsDS5ReplicatedAttributeListTotal attribute can be added to a
replication agreement to list the attributes to exclude from the
total update protocol.  This has the same format as the existing
nsDS5ReplicatedAttributeList attribute.  If a separate total update
attribute list is not specified, the nsDS5ReplicatedAttributeList
value will be used for both incremental and total update.  To use
a total update attribute list, an incremental list must be specified
as well.