make memberof transaction aware and able to be a betxnpostoperation plugin
These changes convert memberof to be able to be a betxnpostoperation
plugin.  It detects from the plugin entry if it is a betxnpostop or a
regular postop (which requires the plugin config entry to be available).
The biggest change was adding a void *txn parameter to almost every function.
Memberof heavily uses recursive calls to descend through nested groups, and
does not pass in the pblock to many of the calls, so the extra txn
parameter was necessary.
Reviewed by: nkinder, nhosoi (Thanks!)