Bug 743966 - Compiler warnings in account usability plugin
This addresses 2 compiler warnings that are seen when compiling the
account usability plug-in.  One issue is that we are passing a 'char *'
to slapi_get_pwpolicy(), which expects a 'Slapi_DN *'.  The other
issue is that we are calling plugin_call_acl_plugin(), which is not
a public function.  This function is wrapped by the public function
slapi_access_allowed(), which we should use instead.