Keep unhashed password psuedo-attribute in the adding entry
Description:
unhashed password pseudo-attribute is necessary for plugins that
handle passwords.  The plugin could be any one including pre/post-
plugin and be-txn-pre/post-plugin.  On the other hand, the pseudo-
attribute should not be in the database.

This patch declares the unhashed password pseudo-attribute as an
operational attribute and allows the pseudo-attribute in an entry
in the memory.  When creating a flat entry by entry2str_ function
to store in the database, it omits the pseudo-attribute.

Reviewed by rmeggins@redhat.com and nkinder@redhat.com (Thank you!!)