Bug 750625 - Fix Coverity (11094) Dereference after null check
https://bugzilla.redhat.com/show_bug.cgi?id=750625

slapd/dse.c (dse_callback_removefromlist)

Bug Description: Passing null variable "filter" to function
"strcmpi_fast", which dereferences it.
If filter and t->filter have a chance to be NULL, the case
should be avoided before passing them to strcasecmp.

Fix Description: adding a check if "filter" and "t->filter"
are NULL or not.  If both are not NULL, they are passed to
strcasecmp.  If either is NULL and another is not, that is
they did not match.