Bug 750625 - Fix Coverity (11055-2) Explicit null dereferenced
https://bugzilla.redhat.com/show_bug.cgi?id=750625

plugins/automember/automember.c (automember_pre_op)

Bug Description: Passing null variable "e" to function
"automember_parse_config_entry", which dereferences it.
Missing a check of the NULL possibility of target entry "e" when
modop is not a supported op type.

Fix Description: If modop is not a supported type, it bails there.