reduce calls to csn_as_string and slapi_log_error
csn_as_string is an expensive operation - we should reduce calls to it
many of these calls are in calls to slapi_log_error at the
REPL log level where we want to print the csn string to the error log
When we call slapi_log_error(REPL,...), any nested functions are called
first, which means that even if the log level is REPL and nothing is logged,
a lot of computation may occur for nothing.  We should protect these calls
using slapi_is_loglevel_set().
Reviewed by: nkinder (Thanks!)