Bug 730387 - attrsyntax - Add slapi_rwlock API and use POSIX rwlocks
This adds a new slapi_rwlock API and uses it throughout the server.
Internally, this API can allow either NSPR or POSIX rwlocks to be
used for different platforms.  This patch makes it use the POSIX
implementation on Linux platforms since the NSPR implementation
does not safely allow re-entrant reader locks to be used.
backported from
commit f9b199ec05f4fd0d3054f05ce6cd6ba4a820d246
Author: Nathan Kinder <nkinder@redhat.com>
Date:   Wed Aug 17 09:20:26 2011 -0700
this patch only addresses the attrsyntax rwlocks, not all of the other
rwlocks in the server