Bug 745259 - Incorrect entryUSN index under high load in              replicated environment
https://bugzilla.redhat.com/show_bug.cgi?id=745259

Bug Description: When replication conflicts occur in replacing
entryusn, dangling entryusn keys pointing the same entryid are
generated in the entryusn index file.  There is no way to clean
them up unless reindexing the entryusn.

Fix Description: When replication conflicts occur in the replace
op and new entry still contains the old value, then the old value
won't be removed from the index file.  Also, if the new value is
not added to the entry, the new value won't be added to the index
file.

The failure is informed to entryusn by by setting LDAP_MOD_IGNORE
in the mod.  Entryusn plugin bepostop_modify uses the info to
determine to increment the entryusn or not.