Ticket #264 - upgrade needs better check for "server is running"
https://fedorahosted.org/389/ticket/264
Resolves: Ticket #264
Bug Description: upgrade needs better check for "server is running"
Reviewed by: nhosoi (Thanks!)
Branch: master
Fix Description: The database upgrade tasks need the server to be offline,
and they need for the upgrade mode to be offline.  These upgrade scripts
check for the presence of the .pid file.  If the file is present, upgrade
assumes the server is running, and skips tasks that require the server to
be offline.  However, under certain conditions (systemd) the .pid files
are left over even after the server is shutdown.  This adds an explicit test
to see if the pid in the pidfile is running.  Note that when pids are
recycled, we could still get a false positive - there could be another
process with the same pid.  In this case, the user will have to manually
remove the .pid files and re-run setup-ds.pl -u.
Platforms tested: RHEL6 x86_64
Flag Day: no
Doc impact: no