Trac Ticket #274 - Reindexing entryrdn fails if                    ancestors are also tombstoned
https://fedorahosted.org/389/ticket/274

Bug description: Inserting/traversing entryrdn fails if a parent
entry is tombstoned and the rdn in the entryrdn index includes
nsuniqueid.

In DIT cn=A,ou=B,o=C, cn=A and ou=B are removed and turned to
tombstone entries.  Both of the 2 representations need to be
supported in the entryrdn.
   nsuniqueid=...,cn=A,ou=B,o=C and
   nsuniqueid=...,cn=A,nsuniqueid=...,ou=B,o=C

Fix description: Support for the second case is added by this patch.
Also, in index_add_mods, code for checking NULL mods is added.