Trac Ticket #275 - Invalid read reported by valgrind
https://fedorahosted.org/389/ticket/275

Fix description: Since the matching rule type could be normalized
and the original string could be freed in filter_normalize_ext,
the type needs to have a duplicated string (bitwise.c, plugin_mr.c).

Filter_ava functions and filter_sub functions in the syntax plugins
need to check if the passed pblock is NULL or not before accessing
it.