Trac Ticket 51 - memory leaks in 389-ds-base-1.2.8.2-1.el5?
https://fedorahosted.org/389/ticket/51

Fix description: Ran valgrind with the MMR+SASL servers and
fixed leaks found in the test.
[plugin/replication/repl5_connection.c]
conn_connect could have overridden conn->ld without releasing
it.  This patch releases it if necessary.
[slapd/dn.c]
If DN normalization fails in slapi_sdn_get_dn, this patch
releases the locally strdup'ed string.
[slapd/modify.c, modutil.c]
DN syntax attribute value is found in mods, it was normalized
and replaced in slapi_mods_insert_at.  It leaked the pre-
noralized value.  Instead, this patch normalizes mods in
do_modify and frees it when the modify is done.
[slapd/operation.c]
modrdn_newsuperior_address.sdn was not release when the modrdn
operaton is done.  This patch adds the release code.