Trac Ticket #27 - SASL/PLAIN binds do not work
https://fedorahosted.org/389/ticket/27

Bug description: ids_sasl_canon_user failed to set "dn: " in front
of the dn string in the output argument out_user.  The dn string
is used in the next session and the corresponding entry was not
found due to the bad dn format (missing "dn: ").

Fix description: This patch adds the proper prefix.