Trac Ticket #84 - 389 Directory Server Unnecessary Checkpoints
https://fedorahosted.org/389/ticket/84

Fix description: txn_checkpoint was always called with DB_FORCE flag.
This patch introduces db_force arg to dblayer_txn_checkpoint and
DB_FORCE is passed only from dblayer_force_checkpoint.

Note: checkpoint_threadmain is one of the BDB housekeeping threads.
It calls txn_checkpoint periodically. The interval is specified in
the ldbm database config:
  dn: cn=config,cn=ldbm database,cn=plugins,cn=config
  nsslapd-db-checkpoint-interval: <second>
Even if DB_FORCE is not set and there is no db modify activities,
as long as checkpoint thread is functioning, some disk IO is
observed due to updating the lock table and mempool to check if
there is any data to flush.