Trac Ticket #26 - Please support setting defaultNamingContext in the rootdse.
https://fedorahosted.org/389/ticket/26

Fix description:
If a config param is set to nsslapd-allowed-to-delete-attrs,
the value is allowed to delete.  nsslapd-defaultnamingcontext
is set to the value, by default.  The config set API is not
designed to allow deleting a param.  Instead, it sets NULL to
represent the deletion.  But it turned out it was not allowed,
either.  This patch allows to the config params set in the
nsslapd-allowed-to-delete-attrs to pass NULL value.