Revert "Ticket #111 - ability to control behavior of modifyTimestamp/modifiersName"
This reverts commit f7b882a8eebf62f15a7802ae4048c8153575d11e.
This is a partial revert.  We still want to keep the new config parameter,
we just don't want to have to change the plugins in any way.  We will
come up with a new mechanism for keeping track of the original requestor
DN, most likely a scheme using Thread Local Storage (TLS).