Revert "Ticket #167 - Mixing transaction and non-transaction plugins can cause deadlock"
This reverts commit 77fdecc3a2bf6d9ce8e545ba0746f619fb14ce84
This is a partial revert.  We still want to keep the code in the plugins
that allows them to be either a regular plugin or a betxn plugin.  All
of the code that dealt with SLAPI_TXN and making sure all internal
operations had SLAPI_TXN set is gone.  We will
come up with a new mechanism for keeping track of the original requestor
DN, most likely a scheme using Thread Local Storage (TLS).